sessions update

+++ bookworm-api/controllers/bookmark.js
@@ -4,6 +4,8 @@ const bookmark = express.Router();
 const Bookmark = require("../models/bookmarkSchema");
 const User = require("../models/users");
 
+const { protect } = require("../middleware/authMiddleware");
+
 bookmark.get("/", (req, res) => {
   Bookmark.find({}, (err, myBookmark) => {
     if (err) {
@@ -14,8 +16,8 @@ bookmark.get("/", (req, res) => {
   });
 });
 
-bookmark.post("/", (req, res) => {
-  Bookmark.create(req.body, (err, newBookmark) => {
+bookmark.post("/", protect, (req, res) => {
+  Bookmark.create(req.body, req.user.id, (err, newBookmark) => {
     if (err) {
       res.status(400).json({ err: err.message });
     } else {
@@ -24,7 +26,16 @@ bookmark.post("/", (req, res) => {
   });
 });
 
-bookmark.delete("/:id", (req, res) => {
+bookmark.delete("/:id", protect, (req, res) => {
+  const user = User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  if (bookmark.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
   Bookmark.findByIdAndDelete(req.params.id, (err, deletedBookmark) => {
     if (err) {
       res.status(400).json({ err: err.message });
@@ -34,7 +45,16 @@ bookmark.delete("/:id", (req, res) => {
   });
 });
 
-bookmark.put("/:id", (req, res) => {
+bookmark.put("/:id", protect, (req, res) => {
+  const user = User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  if (bookmark.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
   Bookmark.findByIdAndUpdate(
     req.params.id,
     req.body,
